+++ learnyounode/emerson-asto/06-mymodule.js
